Let the SDK build the per-key crypto clients
The client composed a key URI by hand and constructed each
CryptographyClient with its own credential and transport, restating what
it had already given the KeyClient. The SDK does this itself:
GetCryptographyClient hands down the parent's credential, options and
pipeline, so the injected transport carries through and no URI is glued
from strings.

Three fields existed only to feed that hand-rolled construction and are
now gone, which is the tell that it was never needed.

It also tightens a contract that was loose by accident. The old string
concat accepted a kid with no version and let Key Vault silently pick the
latest, while this client publishes every kid with its version and the
Vault twin rejects an unversioned kid outright. The kid is now parsed the
same way here, and the tests pass the versioned kid they would really
receive.
