+++ Abblix.Oidc.Server.Azure.UnitTests/StaticTokenCredential.cs
@@ -1,0 +1,39 @@
+// Abblix OIDC Server Library
+// Copyright (c) Abblix LLP. All rights reserved.
+//
+// DISCLAIMER: This software is provided 'as-is', without any express or implied
+// warranty. Use at your own risk. Abblix LLP is not liable for any damages
+// arising from the use of this software.
+//
+// LICENSE RESTRICTIONS: This code may not be modified, copied, or redistributed
+// in any form outside of the official GitHub repository at:
+// https://github.com/Abblix/OIDC.Server. All development and modifications
+// must occur within the official repository and are managed solely by Abblix LLP.
+//
+// Unauthorized use, modification, or distribution of this software is strictly
+// prohibited and may be subject to legal action.
+//
+// For full licensing terms, please visit:
+//
+// https://oidc.abblix.com/license
+//
+// CONTACT: For license inquiries or permissions, contact Abblix LLP at
+// info@abblix.com
+
+using Azure.Core;
+
+namespace Abblix.Oidc.Server.Azure.UnitTests;
+
+/// <summary>
+/// Credential that returns a fixed token without any network call, so the Azure SDK's authentication pipeline
+/// never reaches Entra ID during a test.
+/// </summary>
+internal sealed class StaticTokenCredential : TokenCredential
+{
+    public override AccessToken GetToken(TokenRequestContext requestContext, CancellationToken cancellationToken)
+        => new("stub-token", DateTimeOffset.MaxValue);
+
+    public override ValueTask<AccessToken> GetTokenAsync(
+        TokenRequestContext requestContext, CancellationToken cancellationToken)
+        => ValueTask.FromResult(GetToken(requestContext, cancellationToken));
+}
