+++ Abblix.Oidc.Server.Azure.UnitTests/StubHttpMessageHandler.cs
@@ -1,0 +1,47 @@
+// Abblix OIDC Server Library
+// Copyright (c) Abblix LLP. All rights reserved.
+// 
+// DISCLAIMER: This software is provided 'as-is', without any express or implied
+// warranty. Use at your own risk. Abblix LLP is not liable for any damages
+// arising from the use of this software.
+// 
+// LICENSE RESTRICTIONS: This code may not be modified, copied, or redistributed
+// in any form outside of the official GitHub repository at:
+// https://github.com/Abblix/OIDC.Server. All development and modifications
+// must occur within the official repository and are managed solely by Abblix LLP.
+// 
+// Unauthorized use, modification, or distribution of this software is strictly
+// prohibited and may be subject to legal action.
+// 
+// For full licensing terms, please visit:
+// 
+// https://oidc.abblix.com/license
+// 
+// CONTACT: For license inquiries or permissions, contact Abblix LLP at
+// info@abblix.com
+
+using System.Net;
+using System.Text;
+
+namespace Abblix.Oidc.Server.Azure.UnitTests;
+
+/// <summary>
+/// Records the requests the Azure SDK sends through its transport and returns the canned response the responder
+/// builds, so <see cref="AzureKeyVaultClient"/> is exercised against Key Vault wire shapes without a live vault.
+/// </summary>
+internal sealed class StubHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responder)
+    : HttpMessageHandler
+{
+    public List<HttpRequestMessage> Requests { get; } = [];
+    public HttpRequestMessage? LastRequest => Requests.Count == 0 ? null : Requests[^1];
+
+    protected override Task<HttpResponseMessage> SendAsync(
+        HttpRequestMessage request, CancellationToken cancellationToken)
+    {
+        Requests.Add(request);
+        return Task.FromResult(responder(request));
+    }
+
+    public static HttpResponseMessage Json(HttpStatusCode status, string json)
+        => new(status) { Content = new StringContent(json, Encoding.UTF8, "application/json") };
+}
