+++ Abblix.Oidc.Server.Azure/AzureKeyVaultClient.cs
@@ -47,9 +47,6 @@ namespace Abblix.Oidc.Server.Azure;
 /// </summary>
 public sealed class AzureKeyVaultClient : IKeyCustodian
 {
-    private readonly Uri _vaultUri;
-    private readonly TokenCredential _credential;
-    private readonly HttpClient _httpClient;
     private readonly KeyVault.KeyClient _keyClient;
     private readonly ConcurrentDictionary<string, CryptographyClient> _cryptographyClients = new();
 
@@ -76,11 +73,13 @@ public sealed class AzureKeyVaultClient : IKeyCustodian
     /// <param name="httpClient">The transport for every Key Vault call.</param>
     internal AzureKeyVaultClient(AzureKeyVaultOptions settings, TokenCredential credential, HttpClient httpClient)
     {
-        _vaultUri = new Uri(settings.KeyVaultUri);
-        _credential = credential;
-        _httpClient = httpClient;
+        // The only client this type builds. Everything else it needs comes off this one: the SDK hands the
+        // credential, the options and the pipeline down to the per-key crypto clients, so the injected transport
+        // reaches them without being restated, and no key URI is ever composed by hand here.
         _keyClient = new KeyVault.KeyClient(
-            _vaultUri, _credential, new KeyVault.KeyClientOptions { Transport = new HttpClientTransport(httpClient) });
+            settings.KeyVaultUri,
+            credential,
+            new KeyVault.KeyClientOptions { Transport = new HttpClientTransport(httpClient) });
     }
 
     // Use explicit service-principal credentials from configuration when all three are set; otherwise fall back to
@@ -121,7 +120,7 @@ public sealed class AzureKeyVaultClient : IKeyCustodian
         SigningAlgorithms.ES256 => SignatureAlgorithm.ES256,
         SigningAlgorithms.ES384 => SignatureAlgorithm.ES384,
         SigningAlgorithms.ES512 => SignatureAlgorithm.ES512,
-        
+
         _ => throw new NotSupportedException($"The Azure Key Vault store does not sign '{algorithm}'."),
     };
 
@@ -211,12 +210,41 @@ public sealed class AzureKeyVaultClient : IKeyCustodian
         throw new NotSupportedException($"The Azure Key Vault store does not publish key type '{webKey.KeyType}'.");
     }
 
-    private CryptographyClient GetCryptographyClient(string keyName)
+    /// <summary>
+    /// The crypto client for a key version, cached because building one costs a metadata resolve on first use.
+    /// </summary>
+    /// <param name="keyId">The published <c>kid</c>: the key name and its version, as this client stamped it.</param>
+    /// <remarks>
+    /// The SDK builds the client from the parent <see cref="KeyVault.KeyClient"/>, which hands down its own
+    /// credential, options and pipeline - so the injected transport carries through without being restated, and
+    /// the key's URI is composed by the SDK rather than by string concatenation here.
+    /// </remarks>
+    private CryptographyClient GetCryptographyClient(string keyId)
+        => _cryptographyClients.GetOrAdd(
+            keyId,
+            id =>
+            {
+                var (name, version) = ParseKeyId(id);
+                return _keyClient.GetCryptographyClient(name, version);
+            });
+
+    /// <summary>
+    /// Splits a published <c>kid</c> back into the name and version the SDK addresses a key by.
+    /// </summary>
+    /// <remarks>
+    /// The kid is minted as <c>name/version</c> when the versions are published, and a Key Vault key name cannot
+    /// contain a slash, so the split is unambiguous. A kid without a version is not one this client published.
+    /// </remarks>
+    private static (string Name, string Version) ParseKeyId(string keyId)
     {
-        return _cryptographyClients.GetOrAdd(
-            keyName,
-            name => new CryptographyClient(
-                new Uri($"{_vaultUri}keys/{name}"), _credential,
-                new CryptographyClientOptions { Transport = new HttpClientTransport(_httpClient) }));
+        var separator = keyId.IndexOf('/');
+        if (separator <= 0 || separator == keyId.Length - 1)
+        {
+            throw new InvalidOperationException(
+                $"Malformed external key id '{keyId}': expected '<name>/<version>', which is what this client " +
+                "publishes as the kid of each key version.");
+        }
+
+        return (keyId[..separator], keyId[(separator + 1)..]);
     }
 }
